asfquart/pyproject.toml
[tool.poetry]
name = "asfquart"
version = "0.1.9"
authors = ["ASF Infrastructure <users@infra.apache.org>"]
license = "Apache-2.0"
readme = "README.md"
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3"
]
description = "ASF Quart Framework"

[tool.poetry.dependencies]
python = ">=3.10,<4"
aiohttp = "^3.9.2"
PyYAML = "^6.0.1"
pytest = "7.2.0"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.20.3"
pytest-mock = "^3.10.0"
quart = "^0.19.4"
ezt = "~1.1"
asfpy = "~0.52"
easydict = "~1.13"
exceptiongroup = { version = ">=1.1.0", python = "<3.11" }
watchfiles = "~0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
